feat: Correctly resolve table and column references including database and schema references
